fix: move types into src/types so the generated output can find the path to type def file correctly
